Added support for the return statement
